WARNING client: drop more state on link
The odds this is the wrong solution are way up there. I'm cargo-culting
this from how ::_create() works, because it forces a release of the
parent directory and we need to force a cap flush or the MDS might
evaluate our request with incorrect uid/gid set on the dir. But as I look
at that I find myself completely unclear on the release code, which isn't
documented. Generally we indicate release of stuff we know we need to
drop for the request to go through, but 1) I thought we actually dropped
it, and apparently what actually happens is the MDS just sends cap revokes
before processing (really? so slow), 2) we're unconditionally dropping *parent*
directories when we drop a dentry (I guess because they own it?), 3) these
dentries don't exist anyway.
Is this a hack to deal with the need to have updated directory ownership
state, and so it's the correct solution and I need to get over it? Have
I discovered a generic issue? Something else?

Signed-off-by: Greg Farnum <gfarnum@redhat.com>