mon: Rearrange Paxos::dispatch to be a little cleaner
The Message is actually always a MMonPaxos (which is not a subtype of
PaxosServiceMessage, so that whole thing was wrong!). This lets
us clean up the switch quite a bit and correct the odd casts
going around.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 0c85f8fe68c36575c9a6cd82239a3b91e2c121ae)